requirements.txt
flask==2.0.1
requests==2.26.0
python-dotenv==0.19.0
openai>=1.0.0
python-weather==0.1.0
pillow>=10.0.0
geopy==2.2.0
Werkzeug==2.0.3 